kinder-models: move cube symmetry helpers into dynamic3d utils.py
